feat: add suggested changes, tests, and stdin support
